Revert "Add GSI targets using soong built system image"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/89aa76b1-fe19-4aba-b0fd-220eba7c4040). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ic05a11d40e7bc9c890f004538476f3b20e97be7d
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com