Revert "Add GSI targets using soong built system image"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/fb2e90d8-e541-4b2f-a905-b6742cac6b35). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ie96f509e5c939ae86d82dcffe3cf7ac8efb19be1
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com