Revert "Add AAOS Display Safety triage team"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/f3337974-8f19-4bdf-b3ec-0effade6c29b). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ia3040deef86d39f89ca41a430e47f1b69b8610e8
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com