Revert "Add GSI targets using soong built system image"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/91de6f84-fbfd-4eef-8e73-21d219987d70). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I14c4b14f011313e5288c34034e409e2c41a24cc4
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com