Revert "Fix a race condition in edit monitor"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/18e50d1f-3d10-4a12-87ce-4c35e3e3a2b1). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ibd33a82988d761b471f232ac609b9187ec14685a
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com