Revert "Remove unassigned event log tags feature"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/07b3f9c4-61d7-491e-8876-33f4e8a3b27b). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 383362497
Change-Id: I656f8de8ac348e873e12151a5c436461fd0a24b3
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com