Separate the event-log-tags files per partition
Currently, all event log tags for all modules in the source tree are put into
/system/etc/event-log-tags. This is a problem for incremental soong, because
it needs to analyze all modules in the tree, and treble, because the system
image is now coupled to the other images.

Split up /system/etc/event-log-tags into separate /<partition>/etc/event-log-tags
files. At runtime, all the files will be read and merged, allowing duplicates
because there may be a library with event log tags that's used in two different
binaries on different partitions.

Bug: 382515940
Test: Presubmits
Change-Id: Ib4081d39c3e52858b5e034fc4fb33a5ad633cac9