Revert "Fix broken enable_lz4diff flag"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/adf9bcf1-3122-40bc-ab75-f838f1a904a9). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I541374462d65bb67b1049b68c675b1f1d121e403
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com