Revert "Fix a race condition in edit monitor"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/4728641f-a6c9-4e34-b351-0fc7d9e2b1a2). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I218376f872e6194c771e56155de041fa5e188f48
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com