Revert "Add GSI targets using soong built system image"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/ef57090d-b220-4568-b620-3bec44df4508). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: If9b02d8547eaa4c5b89778d027a45d970119a9eb
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com