Normalize android_info.prop filepath comment in build.prop
build.prop contains a `# from $filepath` comment to log the origin of a
sysprop. This is useful to have, but causes the metadata of
vendor/build.prop to be different between make and soong built
partitions. This CL replaces the absolute filepath with
`android_info.prop` in the build.prop comment.

Bug: 382295862
Test: commented out notice files in build/make/core/Makefile
Test: lunch aosp_cf_x86_64_phone-trunk_staging-userdebug
Test: export BUILD_DATETIME=10
Test: verified that vendor.img built by make and soong have the same
sha

Change-Id: I48f28aae6f6c1d51dc37e9cd19865fe6eb7c8ea9