Revert "Add GSI targets using soong built system image"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/9f336ac0-101a-4338-9c0f-0cc19a3c13d7). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I05cb6b44f22771a5b1d1efa3c1e5b8f642086274
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com