Revert "GSI: build pvmfw"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/bc96230b-337a-41b5-b464-063dbf844f25). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 383995072
Change-Id: I23cea00c8e43e7f5b6e2130ebf12211255ed6f35
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com