Revert "GSI: build pvmfw"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/6d7a3694-df17-4564-a826-3422aad32aac). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 383988648
Bug: 383995687
Change-Id: Ida2260fd4785b09910bbe84045e2557704f9848c
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com