Remove system_linker_config module
The linker.config.pb on the system image has some system-specific
build steps. The system_linker_config module was just one of the
steps, but the rest of the steps depend on installed files, which
are only known to the filesystem modules, so they happen there.
Move this first initial step to the filesystem level as well for
consistency.

Bug: 384091387
Test: Presubmits
Change-Id: Ia27e1b72bd0a003b4043bf3ff089383cff3cd486