Revert "GSI: build pvmfw"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/d6c93cd8-0c6c-4640-aac1-ff3f9d6673be). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 384076807
Change-Id: I453f9f55459742c39ebd247bd12147ad79859162
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com