Revert "GSI: build pvmfw"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/327bb558-a663-494b-a6dc-cd782c8128d5). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 384117466
Change-Id: Iabc9dab96017c372add3ad43125ea0e8e698c97f
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com