Revert "remove StorageInternalReader"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/0ceb22ee-6336-4db9-bf22-3b1528e023cc). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Id34bace55a3057be6637028cb2b8db95b0a68076
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com