Revert "Throw AconfigStorageException instead of AconfigStorageR..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/cec2ae46-7112-471f-a21e-777ea278d340). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I6206b6870a87be00f47c43852e4fc852de1eb59b
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com