Revert "remove StorageInternalReader"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/cec2ae46-7112-471f-a21e-777ea278d340). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I36125f24901a1822e7632823003376dda3de6f8d
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com