Revert "Throw AconfigStorageException instead of AconfigStorageR..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/9d79c478-315d-4a55-8e46-18fbb0069abc). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Iba7600d06e3263567240d0f4f8aadba19827c78d
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com