Revert "Export TARGET_RECOVERY_UI_* variables to Soong"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/c4a3374a-8413-48b3-b852-02c79806fb4d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I6c138f11decb65e0f1715676961a9aae7fa96885
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com