Revert "remove StorageInternalReader"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/bde135ed-b3f5-49f2-9cde-12499de37e7d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ibeffb30ab5c49900e56c351090ba91d2eea9ea08
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com