Revert "remove StorageInternalReader"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/c684163b-9dd9-4c45-8e9f-5780e5d6d54e). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I863bdf817bc97b1caa6a03c5b65b93c3168b5a44
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com