Revert "Throw AconfigStorageException instead of AconfigStorageR..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/b3f03453-40ba-401e-b664-12c74e5ef6ea). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I770674e4d4f49931ca2d9337c3ad26754ef85f3f
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com