Revert "Page size alignment - ignore 32-bit."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/47fd04ab-571e-4cc2-800f-19c9dccd7e60). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I709e3f40a63a39834b4c4baad4d36cfbaf681f64
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com