Revert "Export TARGET_RECOVERY_UI_* variables to Soong"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/4cf4d3a9-f6a9-4ab5-8905-fd04d39580b0). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I8540c43ffde15d81909d1c4f49eafc445f4c74d1
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com