Revert "Return default value if fingerprint doesn't match"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/739b5e00-e702-4484-b783-d03c771112bb). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ieb1fbc6c81e020880f09d23e88c33e3e70a53062
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com