Revert "Extend merge_ota with extra options for signapk.jar"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/2dee037e-df44-4335-acff-546e3e2958c4). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I5fb40133107fa23b0366b8cf5657f37b65d667f0
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com