Revert "Add GSI targets using soong built system image"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/47109a0a-36ec-462f-859f-5e24aa032aa8). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I94f107ce73ccb28da25f407c81d50335758d67e3
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com