Revert "Add modprobe (toolbox) to generic ramdisk"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/20212a43-56c1-4cf7-bd4a-7eb5bf9f98dd). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 389116741
Change-Id: I92175e5bf80e17e5649a83a620a093de0a884cd2
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com